+++ ui/main/game/galactic_war/cards/gwaio_upgrade_advancedtorpedolauncher.js
@@ -4,7 +4,7 @@ define([
   return {
     visible: _.constant(true),
     describe: _.constant(
-      "!LOC:Advanced Torpedo Launcher Upgrade Tech enables the targeting of hover and coastal naval units by the Advanced Torpedo Launcher."
+      "!LOC:Advanced Torpedo Launcher Upgrade Tech enables the targeting of all surface units by the Advanced Torpedo Launcher."
     ),
     summarize: _.constant("!LOC:Advanced Torpedo Launcher Upgrade Tech"),
     icon: _.constant(
@@ -35,9 +35,63 @@ define([
       inventory.addMods([
         {
           file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_tool_weapon.json",
+          path: "spawn_layers",
+          op: "replace",
+          value: "WL_Air",
+        },
+        {
+          file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_tool_weapon.json",
+          path: "target_layers",
+          op: "push",
+          value: ["WL_LandHorizontal"],
+        },
+        {
+          file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_tool_weapon.json",
           path: "exclude_unit_types",
           op: "replace",
           value: "",
+        },
+        {
+          file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_ammo_land.json",
+          path: "flight_layer",
+          op: "replace",
+          value: "Air",
+        },
+        {
+          file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_ammo_land.json",
+          path: "spawn_layers",
+          op: "replace",
+          value: "WL_Air",
+        },
+        {
+          file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_ammo_land.json",
+          path: "cruise_height",
+          op: "replace",
+          value: 75,
+        },
+        {
+          file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_ammo_water.json",
+          path: "flight_layer",
+          op: "replace",
+          value: "Air",
+        },
+        {
+          file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_ammo_water.json",
+          path: "spawn_layers",
+          op: "replace",
+          value: "WL_Air",
+        },
+        {
+          file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_ammo_water.json",
+          path: "cruise_height",
+          op: "replace",
+          value: 75,
+        },
+        {
+          file: "/pa/units/sea/torpedo_launcher_adv/torpedo_launcher_adv_ammo_water.json",
+          path: "initial_velocity",
+          op: "replace",
+          value: 100,
         },
       ]);
     },