+++ ui/main/game/galactic_war/cards/gwaio_upgrade_orbitalfabricationbot.js
@@ -4,7 +4,7 @@ define([
   return {
     visible: _.constant(true),
     describe: _.constant(
-      "!LOC:Orbital Fabrication Bot Upgrade Tech allows the orbital fabricator to reclaim units."
+      "!LOC:Orbital Fabrication Bot Upgrade Tech allows the orbital fabricator to build all basic structures."
     ),
     summarize: _.constant("!LOC:Orbital Fabrication Bot Upgrade Tech"),
     icon: _.constant(
@@ -37,10 +37,11 @@ define([
     buff: function (inventory) {
       inventory.addMods([
         {
-          file: "/pa/units/orbital/orbital_fabrication_bot/orbital_fabrication_bot_build_arm.json",
-          path: "reclaim_types",
-          op: "push",
-          value: "Unit",
+          file: "/pa/units/orbital/orbital_fabrication_bot/orbital_fabrication_bot.json",
+          path: "buildable_types",
+          op: "replace",
+          value:
+            "Land & Structure & Basic | Factory & Basic | FabBuild | FabOrbBuild",
         },
       ]);
     },