+++ ui/main/game/galactic_war/cards/gwaio_upgrade_planetaryradar.js
@@ -1,0 +1,68 @@
+define([
+  "coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/functions.js",
+], function (gwaioFunctions) {
+  return {
+    visible: _.constant(true),
+    describe: _.constant(
+      "!LOC:Planetary Upgrade Tech increases the vision of the planetary radar to match its radar."
+    ),
+    summarize: _.constant("!LOC:Planetary Radar Upgrade Tech"),
+    icon: _.constant(
+      "coui://ui/main/game/galactic_war/gw_play/img/tech/gwc_intelligence_fabrication.png"
+    ),
+    audio: function () {
+      return {
+        found: "/VO/Computer/gw/board_tech_available_efficiency",
+      };
+    },
+    getContext: function (galaxy) {
+      return {
+        totalSize: galaxy.stars().length,
+      };
+    },
+    deal: function (system, context, inventory) {
+      var chance = 0;
+      if (
+        gwaioFunctions.hasUnit(
+          "/pa/units/orbital/deep_space_radar/deep_space_radar.json"
+        ) &&
+        (gwaioFunctions.hasUnit(
+          "/pa/units/air/air_factory_adv/air_factory_adv.json"
+        ) ||
+          inventory.hasCard("gwaio_upgrade_airfactory") ||
+          gwaioFunctions.hasUnit(
+            "/pa/units/land/bot_factory_adv/bot_factory_adv.json"
+          ) ||
+          inventory.hasCard("gwaio_upgrade_botfactory") ||
+          gwaioFunctions.hasUnit(
+            "/pa/units/sea/naval_factory_adv/naval_factory_adv.json"
+          ) ||
+          inventory.hasCard("gwaio_upgrade_navalfactory") ||
+          gwaioFunctions.hasUnit(
+            "/pa/units/land/vehicle_factory_adv/vehicle_factory_adv.json"
+          ) ||
+          inventory.hasCard("gwaio_upgrade_vehiclefactory"))
+      )
+        chance = 60;
+
+      return { chance: chance };
+    },
+    buff: function (inventory) {
+      inventory.addMods([
+        {
+          file: "/pa/units/orbital/deep_space_radar/deep_space_radar.json",
+          path: "recon.observer.items.0.radius",
+          op: "multiply",
+          value: 33.33,
+        },
+        {
+          file: "/pa/units/orbital/deep_space_radar/deep_space_radar.json",
+          path: "recon.observer.items.2.radius",
+          op: "multiply",
+          value: 8.3325,
+        },
+      ]);
+    },
+    dull: function () {},
+  };
+});