+++ ui/main/game/galactic_war/cards/gwaio_upgrade_teleporter.js
@@ -4,7 +4,7 @@ define([
   return {
     visible: _.constant(true),
     describe: _.constant(
-      "!LOC:Teleporter Upgrade Tech removes all energy consumption and efficiency requirements from the interplanetary teleporter."
+      "!LOC:Teleporter Upgrade Tech removes all energy consumption and efficiency requirements from the interplanetary teleporter and makes it invisible to radar."
     ),
     summarize: _.constant("!LOC:Teleporter Upgrade Tech"),
     icon: _.constant(
@@ -40,6 +40,12 @@ define([
           path: "teleporter.energy_demand",
           op: "replace",
           value: 0,
+        },
+        {
+          file: "/pa/units/land/teleporter/teleporter.json",
+          path: "recon.observable.ignore_radar",
+          op: "replace",
+          value: true,
         },
       ]);
     },