+++ ui/mods/com.pa.quitch.gwaioverhaul/shared/functions.js
@@ -44,11 +44,8 @@ define([
         return "coui://ui/main/game/galactic_war/shared/img/red-commander.png";
     },
     quellerAIEnabled: function () {
-      var originSystem = model
-        .game()
-        .galaxy()
-        .stars()
-        [model.game().galaxy().origin()].system();
+      var game = model.game();
+      var originSystem = game.galaxy().stars()[game.galaxy().origin()].system();
       if (originSystem.gwaio && originSystem.gwaio.ai === "Queller") {
         return true;
       }